% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_container.R
\name{get_container}
\alias{get_container}
\title{Get Proteowizard (pwiz) Docker Container}
\usage{
get_container()
}
\description{
Pull the pwiz docker container (\code{chambm/pwiz-skyline-i-agree-to-the-vendor-licenses:latest}) from Docker Hub.
}
